Fix double counting in available driving time calculation
Previous determination of the end of the drive start day is based on the allowed start time instead of the game start time.

For games that start late, the allowed start time may be on the next day.

For back to back games, this lead to double counting between the start day driving and the end day driving.